clean: Move code for better testability
