--- src/main/scala/Result.scala
@@ -1,3 +1,0 @@
-object Result extends Enumeration {
-  val Success, Failure = Value
-}
